[FIX] Wrong path in the sendCSS method leading to the impossibility to be used for a viewpoint.
CSS was included using a relative path and could not be included into a viewpoint, I switched it to an absolute path which solved the issue. I did not test whether the previous inclusion still works (in lists/Item.js) as I am not sure how to display or even generate a signel Item view.