[FIX] Relative path in sendCSS caused the impossibility to be used elsewhere.
CSS was included using a relative path and could not be included into a viewpoint, I switched it to an absolute path which solved the issue. I did not test whether the previous inclusion still works (in lists/Item.js) as I am not sure how to display or even generate a signel Item view.