REFACTORING: Unused code removed.
The code was originally used for importing viewpoints that were exported from Argos v2. 
This code is supposed to import mindmaps in Freemind format.
While potentially interesting, this code hasn’t been used, tested nor documented for the last 8 years.
It is safer to decommission it for now.
An updated version could be reintroduced with feature #3.